requirements.txt
numpy>=1.21.0
pyyaml>=6.0
opencv-python>=4.5.0
pillow>=8.0.0 